Break apart PoolCluster into one class per file
