+++ lib/PoolCluster.js
@@ -1,7 +1,9 @@
-var Pool         = require('./Pool');
-var PoolConfig   = require('./PoolConfig');
-var Util         = require('util');
-var EventEmitter = require('events').EventEmitter;
+var Pool          = require('./Pool');
+var PoolConfig    = require('./PoolConfig');
+var PoolNamespace = require('./PoolNamespace');
+var PoolSelector  = require('./PoolSelector');
+var Util          = require('util');
+var EventEmitter  = require('events').EventEmitter;
 
 module.exports = PoolCluster;
 
@@ -31,7 +33,7 @@ PoolCluster.prototype.of = function(pattern, selector) {
 
   selector = selector || this._defaultSelector;
   selector = selector.toUpperCase();
-  if (typeof Selector[selector] === 'undefined') {
+  if (typeof PoolSelector[selector] === 'undefined') {
     selector = this._defaultSelector;
   }
 
@@ -182,92 +184,4 @@ PoolCluster.prototype._getConnection = function(node, cb) {
 
 PoolCluster.prototype._clearFindCaches = function() {
   this._findCaches = {};
-};
-
-/**
- * PoolNamespace
- */
-function PoolNamespace(cluster, pattern, selector) {
-  this._cluster = cluster;
-  this._pattern = pattern;
-  this._selector = new Selector[selector]();
-}
-
-PoolNamespace.prototype.getConnection = function(cb) {
-  var clusterNode = this._getClusterNode();
-  var cluster     = this._cluster;
-  var namespace   = this;
-
-  if (clusterNode === null) {
-    var err = new Error('Pool does not exist.')
-    err.code = 'POOL_NOEXIST';
-    return cb(err);
-  }
-
-  cluster._getConnection(clusterNode, function(err, connection) {
-    var retry = err && cluster._canRetry
-      && cluster._findNodeIds(namespace._pattern).length !== 0;
-
-    if (retry) {
-      return namespace.getConnection(cb);
-    }
-
-    if (err) {
-      return cb(err);
-    }
-
-    cb(null, connection);
-  });
-};
-
-PoolNamespace.prototype._getClusterNode = function _getClusterNode() {
-  var foundNodeIds = this._cluster._findNodeIds(this._pattern);
-  var nodeId;
-
-  switch (foundNodeIds.length) {
-    case 0:
-      nodeId = null;
-      break;
-    case 1:
-      nodeId = foundNodeIds[0];
-      break;
-    default:
-      nodeId = this._selector(foundNodeIds);
-      break;
-  }
-
-  return nodeId !== null
-    ? this._cluster._getNode(nodeId)
-    : null;
-};
-
-/**
- * Selector
- */
-var Selector = {};
-
-Selector.RR = function () {
-  var index = 0;
-
-  return function(clusterIds) {
-    if (index >= clusterIds.length) {
-      index = 0;
-    }
-
-    var clusterId = clusterIds[index++];
-
-    return clusterId;
-  };
-};
-
-Selector.RANDOM = function () {
-  return function(clusterIds) {
-    return clusterIds[Math.floor(Math.random() * clusterIds.length)];
-  };
-};
-
-Selector.ORDER = function () {
-  return function(clusterIds) {
-    return clusterIds[0];
-  };
 };
